[Refactor/#25] throwble to throwable
