+++ app/src/main/java/com/swyp/firsttodo/presentation/onboarding/OnboardingViewModel.kt
@@ -6,7 +6,7 @@ import androidx.lifecycle.viewModelScope
 import com.swyp.firsttodo.core.base.BaseViewModel
 import com.swyp.firsttodo.core.network.model.ApiError
 import com.swyp.firsttodo.domain.model.Role
-import com.swyp.firsttodo.domain.throwble.ProfileError
+import com.swyp.firsttodo.domain.throwable.ProfileError
 import com.swyp.firsttodo.domain.usecase.user.SaveOnboardingProfile
 import com.swyp.firsttodo.presentation.common.message.ErrorMessage
 import dagger.hilt.android.lifecycle.HiltViewModel
